Give action and CHAIN_PASS_THROUGH classic-free homes
The action decorator moves to @ember/object/-action (re-exported from
the @ember/object barrel), so it no longer shares a module with
EmberObject/computed/observer. CHAIN_PASS_THROUGH moves out of
chain-tags into its own module so tracked's import closure stops pulling
the observer-chain machinery.
